+++ changelog.md
@@ -26,6 +26,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Security
 - Nothing.
 
+## 0.5.0-preview - 2019-12-10
+- [Commits](https://github.com/unity-game-framework/ugf-module-serialize-utf8json/compare/0.4.0-preview...0.5.0-preview)
+- [Milestone](https://github.com/unity-game-framework/ugf-module-serialize-utf8json/milestone/6?closed=1)
+
+### Changed
+- Package dependencies:
+    - `com.ugf.module.serialize`: from `0.2.0-preview` to `0.3.0-preview`.
+    - `com.ugf.utf8json`: from `3.0.0-preview` to `3.2.1-preview`.
+
+### Removed
+- Usage of union formatters.
+
 ## 0.4.0-preview - 2019-12-07
 - [Commits](https://github.com/unity-game-framework/ugf-module-serialize-utf8json/compare/0.3.1-preview...0.4.0-preview)
 - [Milestone](https://github.com/unity-game-framework/ugf-module-serialize-utf8json/milestone/5?closed=1)